+++ app/src/main/java/com/example/campusaccesible/Constants.java
@@ -1,0 +1,7 @@
+package com.example.campusaccesible;
+
+public class Constants {
+
+    public static final String FIREBASE_TAG = "FIREBASE";
+    public static final String COLLECTION_LOCATION = "location";
+}
